`moderately_sensitive` -> `highly_sensitive`
Although the `visit_num` action does not generate patient-level data, it
nevertheless generates unsafe data: data that haven't been subject to
Statistical Disclosure Control. Previously, we marked this data as
`moderately_sensitive` so researchers could view it on -- but not
release it from -- L4. The `make_safe` action makes this unnecessary, so
we mark the `visit_num` action as `highly_sensitive`.